luasocket: variant for lua5.1 and lua5.3
Updating luasocket package that is compatible with, and built against,
lua version 5.3 and 5.1.

I made the following modifications:

Makefile - Set the LUAV environment variable in the build variant.
Set an include path in the lua build so that that lua headers could
be found. Also using the make install target as everything is put
in the correct place automatically for the lua 5.3 build.
The lua 5.1 build still uses the old install method because when
I was testing using the "install" target the lua5.1 interpreter
was unable to locate resources.  So that is why there are
two install methods.

Built and tested on BCM5301X.

Signed-off-by: Colby Whitney <colby.whitney@luxul.com>